Change build location for gh-pages
On running make.jl to build the docs, the output HTML docs are generated in the build folder. Since GitHub pages requires the index.html to be in the master/docs folder, I have added a redirect index.html that redirects to master/docs/build/index.html. This will allow us to build and have the changes be hosted instantly.